requirements.txt
pandas==1.0.3
numpy==1.18.5
streamlit==0.62.0
Pillow==7.2.0
scikit_learn==0.23.1
matplotlib==3.2.2